Op/1

+++ Annexio/Models/Country.cs
@@ -9,11 +9,10 @@ namespace Annexio.Models
         public string Region { get; set; }
         public string Subregion { get; set; }
         public string Capital { get; set; }
-        public int Population { get; set; }
+        public long Population { get; set; }
         public IEnumerable<string> Borders { get; set; }
         public IEnumerable<Currency> Currencies { get; set; }
         public IEnumerable<Language> Languages { get; set; }
-
 
     }
 }
